[bt][l2cap] Remove Wno-conversion flags
Remove Wno-conversion exceptions in l2cap/BUILD.gn and fix a few
conversion errors.

Bug: 58162
Test: fx test //src/connectivity/bluetooth/core/bt-host

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789511
Commit-Queue: Ben Lawson <benlawson@google.com>

+++ pw_bluetooth_sapphire/host/l2cap/channel_configuration.cc
@@ -208,7 +208,7 @@ DynamicByteBuffer ChannelConfiguration::UnknownOption::Encode() const {
   DynamicByteBuffer buffer(size());
   MutablePacketView<ConfigurationOption> option(&buffer, payload_.size());
   option.mutable_header()->type = type_;
-  option.mutable_header()->length = payload_.size();
+  option.mutable_header()->length = static_cast<uint8_t>(payload_.size());
 
   // Raw data is already in little endian
   option.mutable_payload_data().Write(payload_);
